scripts/requirements.txt
# Core dependencies
pydantic>=2.0.0
tiktoken>=0.5.0
unstructured[pdf]>=0.10.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Document processing
pypdf>=3.0.0
python-docx>=0.8.11
python-magic>=0.4.27
chardet>=5.0.0

# NLP and text processing
spacy>=3.5.0
nltk>=3.8.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
mypy>=1.5.0
ruff>=0.1.0
black>=23.0.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# JSON and data handling
orjson>=3.9.0